requirement.txt
asgiref
Django
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
PyJWT
sqlparse
tzdata
